Revert "Feature clément pavie"
