+++ README.md
@@ -5,8 +5,3 @@ Projet TER de L3 informatique de Montpellier de l'année scolaire 2023/2024
 
 A ajouter au TOML
 cargo add csv
-
-
-
-To use the server-client, the server need to be run with : "cargo run --bin server"
-When the server is running, you can launch the client with "cargo run --bin client"
