--- client/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "client"
-version = "0.1.0"
-authors = [""]
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-dns-lookup = "2.0.4"
