--- server/test/.gitignore
@@ -1,4 +1,0 @@
-/target
-**/*.rs.bk
-*.exe
-*.pdb
