--- server/test/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "Rust-client-server"
-version = "0.1.0"
-authors = ["Etienne Gauthier <gaue3012@usherbrooke.ca>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
